Fix decimal numbers not working

+++ calculator.py
@@ -2,12 +2,12 @@ sign = input("What do you want to do? +, -, /, or *: ")
 num1 = input("Enter a number: ")
 num2 = input("Enter a second number: ")
 if sign == "+":
-    result = int(num1) + int(num2)
+    result = float(num1) + float(num2)
 if sign == "-":
-    result = int(num1) - int(num2)
+    result = float(num1) - float(num2)
 if sign == "*":
-    result = int(num1) *  int(num2)
+    result = float(num1) *  float(num2)
 if sign == "/":
-    result = int(num1) / int(num2)
+    result = float(num1) / float(num2)
 
-print(int(result))
+print(result)
